mass rename to (python-)kasa
