+++ .pre-commit-config.yaml
@@ -19,7 +19,6 @@ repos:
   rev: stable
   hooks:
   - id: black
-    language_version: python3.7
 
 - repo: https://github.com/pre-commit/pre-commit-hooks
   rev: v2.3.0
