+++ kasa/tests/fixtures/HS100(US)_1.0.json
