+++ kasa/tests/fixtures/HS105(US)_1.0.json
