+++ kasa/tests/fixtures/HS110(EU)_1.0_real.json
