+++ kasa/tests/fixtures/HS110(EU)_2.0.json
