+++ kasa/tests/fixtures/HS110(US)_1.0.json
