+++ kasa/tests/fixtures/HS200(US)_1.0.json
