+++ kasa/tests/fixtures/HS220(US)_1.0.json
