+++ kasa/tests/fixtures/HS300(US)_1.0.json
