+++ kasa/tests/fixtures/KL120(US)_1.0_real.json
