+++ kasa/tests/fixtures/LB100(US)_1.0.json
