+++ kasa/tests/fixtures/LB120(US)_1.0.json
