+++ kasa/tests/fixtures/LB130(US)_1.0.json
