+++ kasa/tests/newfakes.py
