+++ tox.ini
@@ -17,24 +17,24 @@ deps=
   deprecation
   flake8
 commands=
-    py.test --cov --cov-config=tox.ini pyHS100
+    py.test --cov --cov-config=tox.ini kasa
 
 [testenv:flake8]
 deps=
   flake8
   flake8-docstrings
-commands=flake8 pyHS100
+commands=flake8 kasa
 
 [testenv:typing]
 skip_install=true
 deps=mypy
-commands=mypy --ignore-missing-imports pyHS100
+commands=mypy --ignore-missing-imports kasa
 
 [flake8]
-exclude = .git,.tox,__pycache__,pyHS100/tests/newfakes.py,pyHS100/tests/test_fixtures.py
+exclude = .git,.tox,__pycache__,kasa/tests/newfakes.py,kasa/tests/test_fixtures.py
 max-line-length = 88
 per-file-ignores =
-  pyHS100/tests/*.py:D100,D101,D102,D103,D104
+  kasa/tests/*.py:D100,D101,D102,D103,D104
   setup.py:D100
 ignore = D105, D107, E203, E501, W503
 #ignore = E203, E266, E501, W503, F403, F401
@@ -47,11 +47,11 @@ skip_install = true
 commands = pre-commit run --all-files
 
 [coverage:run]
-source = pyHS100
+source = kasa
 branch = True
 omit =
-  pyHS100/cli.py
-  pyHS100/tests/*
+  kasa/cli.py
+  kasa/tests/*
 
 [coverage:report]
 exclude_lines =
@@ -65,5 +65,5 @@ include_trailing_comma=True
 force_grid_wrap=0
 use_parentheses=True
 line_length=88
-known_first_party=pyHS100
-known_third_party=click,deprecation,pytest,setuptools,voluptuous
+known_first_party=kasa
+known_third_party=click,pytest,setuptools,voluptuous
